[#1675] styling survey activity tabs
had to remove the extra fragment
