--- app/src/main/res/layout/datapoints_fragment.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-        xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:tools="http://schemas.android.com/tools"
-        android:orientation="vertical"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-    <include layout="@layout/tab_layout"/>
-
-    <FrameLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent">
-
-        <androidx.viewpager.widget.ViewPager
-                android:id="@+id/pager"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                tools:context="org.akvo.flow.activity.SurveyActivity">
-        </androidx.viewpager.widget.ViewPager>
-
-        <View
-                android:layout_height="6dp"
-                android:layout_width="match_parent"
-                android:background="@drawable/shadow_background"/>
-    </FrameLayout>
-</LinearLayout>
-
