+++ app/src/main/res/layout/survey_activity_content.xml
@@ -1,31 +1,26 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.coordinatorlayout.widget.CoordinatorLayout
-        xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        xmlns:tools="http://schemas.android.com/tools"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:id="@+id/survey_root_layout"
-        tools:context="org.akvo.flow.activity.SurveyActivity">
-
-    <include layout="@layout/app_bar_layout"/>
-
-    <FrameLayout
-            android:layout_marginTop="?actionBarSize"
-            android:id="@+id/content_frame"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"/>
+<androidx.coordinatorlayout.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/survey_root_layout"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context="org.akvo.flow.activity.SurveyActivity">
 
     <com.google.android.material.floatingactionbutton.FloatingActionButton
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:clickable="true"
-            android:layout_marginBottom="16dp"
-            android:layout_marginRight="16dp"
-            android:layout_marginEnd="16dp"
-            android:layout_gravity="end|bottom"
-            android:visibility="gone"
-            tools:visibility="visible"
-            app:srcCompat="@drawable/ic_action_new"
-            android:id="@+id/add_data_point_fab"/>
+        android:id="@+id/add_data_point_fab"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="end|bottom"
+        android:layout_marginEnd="16dp"
+        android:layout_marginRight="16dp"
+        android:layout_marginBottom="16dp"
+        android:clickable="true"
+        android:visibility="gone"
+        app:srcCompat="@drawable/ic_action_new"
+        tools:visibility="visible" />
+
+    <include layout="@layout/datapoint_content" />
+
+    <include layout="@layout/app_bar_layout_tabs" />
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
