+++ uicomponents/src/main/res/layout/app_bar_layout_tabs.xml
@@ -9,7 +9,6 @@
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/toolbar"
         android:layout_width="match_parent"
-        style="@style/Widget.MaterialComponents.Toolbar"
         android:layout_height="?attr/actionBarSize"
         android:background="?attr/colorPrimary"
         app:popupTheme="@style/FlowAppTheme.PopupOverlay" />
